+++ backend/src/main/java/com/weichat/finance/service/PayOrderQueryLogService.java
@@ -1,0 +1,13 @@
+package com.weichat.finance.service;
+
+import com.baomidou.mybatisplus.extension.service.IService;
+import com.weichat.finance.entity.PayOrderQueryLog;
+
+/**
+ * 定时查单日志 Service。
+ *
+ * @author panhw
+ * @since 2026-09-14
+ */
+public interface PayOrderQueryLogService extends IService<PayOrderQueryLog> {
+}